+++ ui/src/dashboards/components/Dashboard.tsx
@@ -13,7 +13,6 @@ import {ErrorHandling} from 'src/shared/decorators/errors'
 interface Props {
   dashboard: Dashboard
   timeRange: TimeRange
-  autoRefresh: number
   manualRefresh: number
   inPresentationMode: boolean
   inView: (cell: Cell) => boolean
@@ -34,7 +33,6 @@ class DashboardComponent extends PureComponent<Props> {
       onZoom,
       dashboard,
       timeRange,
-      autoRefresh,
       manualRefresh,
       onDeleteCell,
       onCloneCell,
@@ -58,7 +56,6 @@ class DashboardComponent extends PureComponent<Props> {
             <Cells
               onZoom={onZoom}
               timeRange={timeRange}
-              autoRefresh={autoRefresh}
               manualRefresh={manualRefresh}
               cells={dashboard.cells}
               onCloneCell={onCloneCell}