+++ ui/src/timeMachine/reducers/index.ts
@@ -14,7 +14,7 @@ import {
 } from 'src/timeMachine/constants'
 
 // Types
-import {TimeRange, View} from 'src/types'
+import {TimeRange, View, AutoRefresh} from 'src/types'
 import {
   ViewType,
   DashboardDraftQuery,
@@ -28,6 +28,7 @@ import {Action} from 'src/timeMachine/actions'
 import {TimeMachineTab} from 'src/types/timeMachine'
 import {RemoteDataState} from 'src/types'
 import {Color} from 'src/types/colors'
+import {AUTOREFRESH_DEFAULT} from 'src/shared/constants'
 
 interface QueryBuilderState {
   buckets: string[]
@@ -53,6 +54,7 @@ interface QueryResultsState {
 export interface TimeMachineState {
   view: QueryView
   timeRange: TimeRange
+  autoRefresh: AutoRefresh
   draftQueries: DashboardDraftQuery[]
   isViewingRawData: boolean
   activeTab: TimeMachineTab
@@ -70,6 +72,7 @@ export interface TimeMachinesState {
 
 export const initialStateHelper = (): TimeMachineState => ({
   timeRange: {lower: 'now() - 1h'},
+  autoRefresh: AUTOREFRESH_DEFAULT,
   view: createView(),
   draftQueries: [{...defaultViewQuery(), hidden: false}],
   isViewingRawData: false,
@@ -172,6 +175,14 @@ export const timeMachineReducer = (
     case 'SET_TIME_RANGE': {
       return produce(state, draftState => {
         draftState.timeRange = action.payload.timeRange
+
+        buildAllQueries(draftState)
+      })
+    }
+
+    case 'SET_AUTO_REFRESH': {
+      return produce(state, draftState => {
+        draftState.autoRefresh = action.payload.autoRefresh
 
         buildAllQueries(draftState)
       })